requirements.txt
python-dotenv
flask
requests
streamlit
googlesearch-python
google-generativeai
beautifulsoup4
sentence-transformers